Issue #380 - Alpaca Crypto Data API (v1-beta1) implementation (RESTFul API part).
(cherry picked from commit f7398e7414600f023b54ba1e66c93b29a0f8c23a)

+++ Alpaca.Markets/Messages/JsonHistoricalBar.cs
@@ -25,7 +25,7 @@ namespace Alpaca.Markets
         public Decimal High { get; set; }
 
         [JsonProperty(PropertyName = "v", Required = Required.Always)]
-        public UInt64 Volume { get; set; }
+        public Decimal Volume { get; set; }
 
         [JsonProperty(PropertyName = "t", Required = Required.Always)]
         public DateTime TimeUtc { get; set; }
